+++ app/Http/Controllers/SubjectController.php
@@ -42,12 +42,13 @@ class SubjectController extends Controller
 
         Subject::create($data);
 
-        Session::flash('message', 'Tạo môn học thành công');
+        Session::flash('create', 'Tạo môn học thành công');
 
         return back();
     }
 
     public function edit(Subject $subject){
+
         $this->authorize('update', $subject);
         
         return view('admin.subject.edit', [
@@ -56,6 +57,7 @@ class SubjectController extends Controller
     }
 
     public function update(Subject $subject){
+
         $data = request()->validate([
             /**
              * không trùng lặp và bỏ qua tên môn cần update
@@ -73,7 +75,7 @@ class SubjectController extends Controller
 
         $subject->update($data);
 
-        Session::flash('updated-message', 'Thông tin môn học được sửa thành công');
+        Session::flash('update', 'Thông tin môn học được sửa thành công');
 
         return redirect()->route('subjects.index');
     }
@@ -82,9 +84,11 @@ class SubjectController extends Controller
 
         $subject = Subject::findOrFail($request->subject_id);
 
+        $this->authorize('delete', $subject);
+
         $subject->delete();
 
-        Session::flash('destroy-message', 'Môn học đã được xóa thành công');
+        Session::flash('destroy', 'Môn học đã được xóa thành công');
 
         return back();
     }
